# docs/swagger-doc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Groccery API documentation",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://groccery-api.onrender.com/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:4000/api",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/orders": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Order"],
        "summary": "Get a list of order items",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "count": { "type": "number" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Order" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "security": [{ "xAuth": [] }],
        "tags": ["Order"],
        "summary": "Create a new order item",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "product": {
                    "type": "string",
                    "default": "6370f340cf88e3ec4c1502a8"
                  },
                  "quantity": { "type": "number", "default": "5" },
                  "totalPrice": { "type": "number", "default": "100" },
                  "address": {
                    "type": "string",
                    "default": "63d90377a373a0aea895f528"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Order created successfully."
                    },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Order" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },

          "422": {
            "description": "Validaton error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/orders/{id}": {
      "get": {
        "security": [{ "xAuth": [] }],
        "tags": ["Order"],
        "summary": "Get a order item with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Order ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "security": [{ "xAuth": [] }],
        "tags": ["Order"],
        "summary": "Delete a order with specified ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Order ID",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "default": "success" },
                    "message": {
                      "type": "string",
                      "default": "Order deleted successfully."
                    },
                    "data": {
                      "type": "object",
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Token error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "xAuth": {
        "type": "apiKey",
        "name": "x-token",
        "in": "header"
      }
    },
    "schemas": {
...
      "Order": {
        "properties": {
          "user": { "type": "string" },
          "product": { "type": "string" },
          "quantity": { "type": "number" },
          "expectedDelivery": { "type": "string", "format": "date-time" },
          "totalPrice": { "type": "number" },
          "deliveryStatus": { "type": "string", "enum": ["Array"] },
          "address": { "type": "string" },
          "_id": { "type": "string" }
        }
      },
      "Error": {
        "properties": {
          "status": { "type": "string", "default": "error" },
          "message": { "type": "string" }
        }
      }
    }
  }
}